requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
docx2pdf==0.1.8
et_xmlfile==2.0.0
idna==3.11
lxml==6.0.2
openpyxl==3.1.5
packaging==26.0
pikepdf==10.2.0
pillow==12.1.0
psutil==7.2.1
pycparser==3.0
pycryptodome==3.23.0
PyMuPDF==1.26.7
PyQt6==6.10.2
PyQt6-Qt6==6.10.1
PyQt6_sip==13.11.0
PyKCS11>=1.5.0
python-dotenv==1.2.1
pywin32==311
endesive>=2.17.0
PyYAML==6.0.3
requests==2.32.5
tqdm==4.67.1
urllib3==2.6.3
wrapt==2.0.1
